fix(favicon + stripe): correct favicon path for production and improve Stripe API debugging
