+++ index.html
@@ -18,15 +18,15 @@
 
     <link rel="manifest" href="/site.manifest" />
 
-    <link rel="apple-touch-icon" sizes="180x180" href="/src/favicons/apple-touch-icon.png" />
-    <link rel="icon" type="image/png" sizes="48x48" href="/src/favicons/favicon-48x48.png" />
-    <link rel="icon" type="image/png" sizes="32x32" href="/src/favicons/favicon-32x32.png" />
-    <link rel="icon" type="image/png" sizes="16x16" href="/src/favicons/favicon-16x16.png" />
-    <link rel="icon" type="image/png" sizes="192x192" href="/src/favicons/android-chrome-192x192.png" />
-    <link rel="icon" type="image/png" sizes="512x512" href="/src/favicons/android-chrome-512x512.png" />
-    <link rel="icon" type="image/x-icon" href="/src/favicons/favicon.ico" />
-    <link rel="shortcut icon" href="/src/favicons/favicon.ico" />
-    <link rel="mask-icon" href="/src/favicons/safari-pinned-tab.svg" color="#1C4ED8">
+    <link rel="apple-touch-icon" sizes="180x180" href="/favicons/apple-touch-icon.png" />
+    <link rel="icon" type="image/png" sizes="48x48" href="/favicons/favicon-48x48.png" />
+    <link rel="icon" type="image/png" sizes="32x32" href="/favicons/favicon-32x32.png" />
+    <link rel="icon" type="image/png" sizes="16x16" href="/favicons/favicon-16x16.png" />
+    <link rel="icon" type="image/png" sizes="192x192" href="/favicons/android-chrome-192x192.png" />
+    <link rel="icon" type="image/png" sizes="512x512" href="/favicons/android-chrome-512x512.png" />
+    <link rel="icon" type="image/x-icon" href="/favicons/favicon.ico" />
+    <link rel="shortcut icon" href="/favicons/favicon.ico" />
+    <link rel="mask-icon" href="/favicons/safari-pinned-tab.svg" color="#1C4ED8">
 
     <meta property="og:image" content="/images/Single_Student_Screenshot.avif" />
     <meta name="twitter:card" content="/images/Single_Student_Screenshot.avif" />
